chore: remove `AnnotatableFileDiff` leftovers, rename file
